refactor: move helpers into /lib/api
