+++ imports/plugins/included/shipping-rates/server/no-meteor/util/attributeDenyCheck.js
@@ -1,4 +1,4 @@
-import { operators, propertyTypes } from "./helpers";
+import { operators, propertyTypes } from "/lib/api";
 
 /**
  * @summary Filter shipping methods based on per method deny attribute restrictions
