+++ lib/api/helpers.js
@@ -305,3 +305,32 @@ export const ValidCVV = Match.Where((x) => /^[0-9]{3,4}$/.test(x));
 
 // Email regex
 export const ValidEmail = (x) => /(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])/.test(x); // eslint-disable-line max-len
+
+
+/**
+ * @name operators
+ * @constant
+ * @memberof Helpers
+ * @summary Checks matching via various operators
+ * @return {Boolean} the result of the check
+ */
+export const operators = {
+  eq(varA, varB) { return varA === varB; },
+  gt(varA, varB) { return varA > varB; },
+  lt(varA, varB) { return varA < varB; },
+  ne(varA, varB) { return varA !== varB; }
+};
+
+/**
+ * @name propertyTypes
+ * @constant
+ * @memberof Helpers
+ * @summary Checks matching via various operators
+ * @return {Boolean} the result of the check
+ */
+export const propertyTypes = {
+  bool(varA) { return varA.trim().toLowerCase() === "true"; },
+  float(varA) { return parseFloat(varA); },
+  int(varA) { return parseInt(varA, 10); },
+  string(varA) { return varA; }
+};
